pyproject.toml
[tool.poetry]
name = "cursachdjango"
version = "0.1.0"
description = ""
authors = ["EvgeniyRibakov <evgenijrybakov01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = "5.3.6"
redis = "5.0.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
django-allauth = "^65.7.0"
django-celery-beat = "^2.8.0"
pillow = "^11.2.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
